take out edit circle
